pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-ai-framework"
version = "0.1.0"
description = "A universal MCP-based AI framework for building intelligent applications with modular tools"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langchain-core>=0.3.0",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]
vector = [
    "pymilvus>=2.3.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
]
ai = [
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "pandas>=2.0.0",
]
langchain = [
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-experimental>=0.3.0",
    "langsmith>=0.1.0",
    "langgraph-checkpoint>=1.0.0",
    "langgraph-checkpoint-sqlite>=1.0.0",
]
ui = [
    "streamlit>=1.28.0",
    "gradio>=4.0.0",
]
all = [
    "pymilvus>=2.3.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "pandas>=2.0.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-experimental>=0.3.0",
    "langsmith>=0.1.0",
    "langgraph-checkpoint>=1.0.0",
    "langgraph-checkpoint-sqlite>=1.0.0",
    "streamlit>=1.28.0",
    "gradio>=4.0.0",
]

[project.scripts]
mcp-framework = "mcp_framework.cli:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 